+++ e3nn_jax/_activation.py
@@ -4,7 +4,7 @@ import jax
 import jax.numpy as jnp
 
 from e3nn_jax import Irreps, IrrepsArray
-from e3nn_jax.util.decorators import overload_for_irreps_without_data
+from e3nn_jax.util.decorators import overload_for_irreps_without_array
 
 
 def normalize_function(phi):
@@ -41,7 +41,7 @@ def is_zero_in_zero(phi):
         return jnp.allclose(phi(jnp.array(0.0)), 0.0)
 
 
-@overload_for_irreps_without_data(irrepsdata_argnums=[0])
+@overload_for_irreps_without_array(irrepsarray_argnums=[0])
 def scalar_activation(input: IrrepsArray, acts: List[Optional[Callable[[float], float]]]) -> IrrepsArray:
     assert isinstance(input, IrrepsArray)
 
@@ -70,7 +70,7 @@ def scalar_activation(input: IrrepsArray, acts: List[Optional[Callable[[float], 
                 if is_zero_in_zero(act):
                     list.append(None)
                 else:
-                    list.append(act(jnp.ones(input.shape + (mul, 1))))
+                    list.append(act(jnp.ones(input.shape[:-1] + (mul, 1))))
             else:
                 list.append(act(x))
         else:
@@ -84,7 +84,7 @@ def scalar_activation(input: IrrepsArray, acts: List[Optional[Callable[[float], 
         contiguous = input.contiguous if acts[0] is None else normalize_function(acts[0])(input.contiguous)
         return IrrepsArray(irreps=irreps_out, contiguous=contiguous, list=list)
 
-    return IrrepsArray.from_list(irreps_out, list, input.shape)
+    return IrrepsArray.from_list(irreps_out, list, input.shape[:-1])
 
 
 # TODO remove this class and follow the same pattern as scalar_activation