+++ e3nn_jax/experimental/transformer.py
@@ -101,11 +101,11 @@ class Transformer(hk.Module):
             edge_dst (array of int32): destination index of the edges
             edge_scalar_attr (array of float): scalar attributes of the edges (typically given by ``soft_one_hot_linspace``)
             edge_weight_cutoff (array of float): cutoff weight for the edges (typically given by ``sus``)
-            edge_attr (IrrepsData): attributes of the edges (typically given by ``spherical_harmonics``)
-            node_f (IrrepsData): features of the nodes
+            edge_attr (IrrepsArray): attributes of the edges (typically given by ``spherical_harmonics``)
+            node_f (IrrepsArray): features of the nodes
 
         Returns:
-            IrrepsData: output features of the nodes
+            IrrepsArray: output features of the nodes
         """
         edge_src_feat = jax.tree_util.tree_map(lambda x: x[edge_src], node_feat)
         edge_dst_feat = jax.tree_util.tree_map(lambda x: x[edge_dst], node_feat)
@@ -127,11 +127,11 @@ class Transformer(hk.Module):
         z = jnp.where(z == 0.0, 1.0, z)
         alpha = exp / z[edge_dst]  # array[edge, head]
 
-        edge_v = edge_v.factor_mul_to_last_axis(self.num_heads)  # IrrepsData[edge, head, irreps_out]
-        edge_v = edge_v * jnp.sqrt(jax.nn.relu(alpha))  # IrrepsData[edge, head, irreps_out]
-        edge_v = edge_v.repeat_mul_by_last_axis()  # IrrepsData[edge, irreps_out]
+        edge_v = edge_v.factor_mul_to_last_axis(self.num_heads)  # IrrepsArray[edge, head, irreps_out]
+        edge_v = edge_v * jnp.sqrt(jax.nn.relu(alpha))  # IrrepsArray[edge, head, irreps_out]
+        edge_v = edge_v.repeat_mul_by_last_axis()  # IrrepsArray[edge, irreps_out]
 
         node_out = jax.tree_util.tree_map(
             lambda x: index_add(edge_dst, x, node_feat.shape[0]), edge_v
-        )  # IrrepsData[node, irreps_out]
-        return jax.vmap(Linear(self.irreps_node_output))(node_out)  # IrrepsData[edge, head, irreps_out]
+        )  # IrrepsArray[node, irreps_out]
+        return jax.vmap(Linear(self.irreps_node_output))(node_out)  # IrrepsArray[edge, head, irreps_out]